feat: focus inputs via query param
feat: add fieldRef to fields in GenericInput

refactor: custom inputs to be focussed

docs: add documentation for useFieldInputFocus hook

feat: work with CM specific components

feat: auto-open repeatable components

feat: make wyiwyg editor focusable

chore: fix other test suites

test: add generic input tests for auto-focussing

chore: add comment for not supported

Update packages/core/admin/admin/src/content-manager/components/RepeatableComponent/index.js
