+++ packages/core/admin/admin/src/content-manager/components/DynamicZone/components/DynamicZoneLabel.js
@@ -29,7 +29,7 @@ export const DynamicZoneLabel = ({
         paddingBottom={3}
         paddingRight={4}
         paddingLeft={4}
-        borderRadius={26}
+        borderRadius="26px"
         background="neutral0"
         shadow="filterShadow"
         color="neutral500"
