--- packages/core/helper-plugin/src/hooks/useQuery.js
@@ -1,7 +1,0 @@
-import { useLocation } from 'react-router-dom';
-
-const useQuery = () => {
-  return new URLSearchParams(useLocation().search);
-};
-
-export { useQuery };
